Refactor trace_processor such that TIDs and PIDs are 64-bit (#1750)
This is to accomodate for OSes which have greater sized TIDs and PIDs.

Bug: 421994886

+++ src/trace_processor/importers/common/process_tracker.cc
@@ -69,7 +69,7 @@ ProcessTracker::ProcessTracker(TraceProcessorContext* context)
 ProcessTracker::~ProcessTracker() = default;
 
 UniqueTid ProcessTracker::StartNewThread(std::optional<int64_t> timestamp,
-                                         uint32_t tid) {
+                                         int64_t tid) {
   tables::ThreadTable::Row row;
   row.tid = tid;
   row.start_ts = timestamp;
@@ -90,7 +90,7 @@ UniqueTid ProcessTracker::StartNewThread(std::optional<int64_t> timestamp,
   return new_utid;
 }
 
-void ProcessTracker::EndThread(int64_t timestamp, uint32_t tid) {
+void ProcessTracker::EndThread(int64_t timestamp, int64_t tid) {
   auto& thread_table = *context_->storage->mutable_thread_table();
   auto& process_table = *context_->storage->mutable_process_table();
 
@@ -132,7 +132,7 @@ void ProcessTracker::EndThread(int64_t timestamp, uint32_t tid) {
   pids_.Erase(tid);
 }
 
-std::optional<UniqueTid> ProcessTracker::GetThreadOrNull(uint32_t tid) {
+std::optional<UniqueTid> ProcessTracker::GetThreadOrNull(int64_t tid) {
   auto opt_utid = GetThreadOrNull(tid, std::nullopt);
   if (!opt_utid)
     return std::nullopt;
@@ -152,12 +152,12 @@ std::optional<UniqueTid> ProcessTracker::GetThreadOrNull(uint32_t tid) {
   return utid;
 }
 
-UniqueTid ProcessTracker::GetOrCreateThread(uint32_t tid) {
+UniqueTid ProcessTracker::GetOrCreateThread(int64_t tid) {
   auto utid = GetThreadOrNull(tid);
   return utid ? *utid : StartNewThread(std::nullopt, tid);
 }
 
-UniqueTid ProcessTracker::UpdateThreadName(uint32_t tid,
+UniqueTid ProcessTracker::UpdateThreadName(int64_t tid,
                                            StringId thread_name_id,
                                            ThreadNamePriority priority) {
   auto utid = GetOrCreateThread(tid);
@@ -206,14 +206,14 @@ bool ProcessTracker::IsThreadAlive(UniqueTid utid) {
     return false;
 
   // If the process has been replaced in |pids_|, this thread is dead.
-  uint32_t current_pid = prr.pid();
+  int64_t current_pid = prr.pid();
   auto* pid_it = pids_.Find(current_pid);
   return !pid_it || *pid_it == current_upid;
 }
 
 std::optional<UniqueTid> ProcessTracker::GetThreadOrNull(
-    uint32_t tid,
-    std::optional<uint32_t> pid) {
+    int64_t tid,
+    std::optional<int64_t> pid) {
   auto& threads = *context_->storage->mutable_thread_table();
   auto& processes = *context_->storage->mutable_process_table();
 
@@ -243,14 +243,14 @@ std::optional<UniqueTid> ProcessTracker::GetThreadOrNull(
 
     // We found a thread that matches both the tid and its parent pid.
     auto prr = processes[*opt_current_upid];
-    uint32_t current_pid = prr.pid();
+    int64_t current_pid = prr.pid();
     if (!pid || current_pid == *pid)
       return current_utid;
   }
   return std::nullopt;
 }
 
-UniqueTid ProcessTracker::UpdateThread(uint32_t tid, uint32_t pid) {
+UniqueTid ProcessTracker::UpdateThread(int64_t tid, int64_t pid) {
   auto& thread_table = *context_->storage->mutable_thread_table();
 
   // Try looking for a thread that matches both tid and thread group id (pid).
@@ -271,19 +271,19 @@ UniqueTid ProcessTracker::UpdateThread(uint32_t tid, uint32_t pid) {
   return utid;
 }
 
-void ProcessTracker::UpdateTrustedPid(uint32_t trusted_pid, uint64_t uuid) {
+void ProcessTracker::UpdateTrustedPid(int64_t trusted_pid, uint64_t uuid) {
   trusted_pids_[uuid] = trusted_pid;
 }
 
-std::optional<uint32_t> ProcessTracker::GetTrustedPid(uint64_t uuid) {
+std::optional<int64_t> ProcessTracker::GetTrustedPid(uint64_t uuid) {
   if (trusted_pids_.find(uuid) == trusted_pids_.end())
     return std::nullopt;
   return trusted_pids_[uuid];
 }
 
-std::optional<uint32_t> ProcessTracker::ResolveNamespacedTid(
-    uint32_t root_level_pid,
-    uint32_t tid) {
+std::optional<int64_t> ProcessTracker::ResolveNamespacedTid(
+    int64_t root_level_pid,
+    int64_t tid) {
   if (root_level_pid <= 0)  // Not a valid pid.
     return std::nullopt;
 
@@ -314,8 +314,8 @@ std::optional<uint32_t> ProcessTracker::ResolveNamespacedTid(
 }
 
 UniquePid ProcessTracker::StartNewProcess(std::optional<int64_t> timestamp,
-                                          std::optional<uint32_t> parent_tid,
-                                          uint32_t pid,
+                                          std::optional<int64_t> parent_tid,
+                                          int64_t pid,
                                           StringId main_thread_name,
                                           ThreadNamePriority priority) {
   pids_.Erase(pid);
@@ -357,8 +357,8 @@ UniquePid ProcessTracker::StartNewProcess(std::optional<int64_t> timestamp,
   return upid;
 }
 
-UniquePid ProcessTracker::SetProcessMetadata(uint32_t pid,
-                                             std::optional<uint32_t> ppid,
+UniquePid ProcessTracker::SetProcessMetadata(int64_t pid,
+                                             std::optional<int64_t> ppid,
                                              base::StringView name,
                                              base::StringView cmdline) {
   std::optional<UniquePid> pupid;
@@ -417,7 +417,7 @@ void ProcessTracker::SetStartTsIfUnset(UniquePid upid,
 }
 
 void ProcessTracker::UpdateThreadNameAndMaybeProcessName(
-    uint32_t tid,
+    int64_t tid,
     StringId thread_name,
     ThreadNamePriority priority) {
   auto& tt = *context_->storage->mutable_thread_table();
@@ -436,7 +436,7 @@ void ProcessTracker::UpdateThreadNameAndMaybeProcessName(
   }
 }
 
-UniquePid ProcessTracker::GetOrCreateProcess(uint32_t pid) {
+UniquePid ProcessTracker::GetOrCreateProcess(int64_t pid) {
   auto& process_table = *context_->storage->mutable_process_table();
 
   // If the insertion succeeds, we'll fill the upid below.
@@ -488,8 +488,8 @@ void ProcessTracker::AssociateThreads(UniqueTid utid1, UniqueTid utid2) {
 
   if (opt_upid1.has_value() && opt_upid1 != opt_upid2) {
     // Cannot associate two threads that belong to two different processes.
-    PERFETTO_ELOG("Process tracker failure. Cannot associate threads %u, %u",
-                  rr1.tid(), rr2.tid());
+    PERFETTO_ELOG("Process tracker failure. Cannot associate threads %ld, %ld",
+                  static_cast<long>(rr1.tid()), static_cast<long>(rr2.tid()));
     context_->storage->IncrementStats(stats::process_tracker_errors);
     return;
   }
@@ -604,14 +604,14 @@ void ProcessTracker::NotifyEndOfFile() {
   namespaced_processes_.clear();
 }
 
-void ProcessTracker::UpdateNamespacedProcess(uint32_t pid,
-                                             std::vector<uint32_t> nspid) {
+void ProcessTracker::UpdateNamespacedProcess(int64_t pid,
+                                             std::vector<int64_t> nspid) {
   namespaced_processes_[pid] = {pid, std::move(nspid), {}};
 }
 
-void ProcessTracker::UpdateNamespacedThread(uint32_t pid,
-                                            uint32_t tid,
-                                            std::vector<uint32_t> nstid) {
+void ProcessTracker::UpdateNamespacedThread(int64_t pid,
+                                            int64_t tid,
+                                            std::vector<int64_t> nstid) {
   PERFETTO_DCHECK(namespaced_processes_.find(pid) !=
                   namespaced_processes_.end());
   auto& process = namespaced_processes_[pid];
